refactor(dashboard): extract audit-log text helpers to lib
react-refresh/only-export-components fails on .tsx files that export
both React components and plain functions — HMR can't draw the
component boundary correctly. Moves getActorLabel, renderVerb,
formatAuditAction (and their internal helpers getResourceLabel,
endpointHost, describeOtelForwardingUpsert, describeToolsetUpdate)
into a new pure-TS lib at @/lib/audit-log-format. OrgAuditLogs.tsx
and OrgHome.tsx import them from the new home.

Fixes client-build-lint failure on the PR.
